pyproject.toml
[project]
name = "stickerbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "uvicorn>=0.42.0",
]
